May need multiple calls to SSL_read.
Otherwise, undrained data can be sat there. We don't get another
event from the socket, and so don't read it until the next data
arrives or the socket is closed. This could cause hangs when the
data was needed to produce a response in order for the interaction
to continue.