Use Supply.Channel.Supply idiom over .schedule-on
The latter can cause disordering, which is probably to be considered
a Rakudo bug. The goal was to make sure that the work is performed
off the emitting thread, so we don't block under lock. By calling
.Channel on the Supply, it is tapped and its content put in a Channel.
Then doing .Supply on that handle the values asynchronously while
also maintaining ordering. `.schedule-on` can be implemented in a
more efficient way than that, but until it's fixed in Rakudo to
maintain order of messages then this is a better way to go.

This fixes occasional corruption of received data, and one of the
causes of #19.