Fix occassional deadlock due to DESTROY
DESTROY can run at any time. This means at some point it will run at
a really unfortunate time. It could, for example, run when attaching
a `.then` to a `Promise`, trying to acquire the lock for libssl to do
the cleanup. At the same time, another thread might have the libssl
lock and want to keep or break that same Promise. This leads to a
circular wait and the program deadlocks. This caused thread-stress.t
to hang occasionally.

Fix it by doing the cleanup asycnrhonously, in the pool, when no locks
are being held. But only schedule that to happen if we really need to
do cleanup, so something well-behaved that closes sockets in an
orderly fashion will not even need to do such scheduling.