Only set up context once per listen
Instead of per accepted connection. Inspired by inclusion of similar
in a PR by s0me0ne-unkn0wn++, but this moves it out of the `whenever`
while still leaving it in the `supply` block. That still means we do
not have to do it per connection, (so we get the same win), but gives
us the chance to do the context cleanup upon `CLOSE` (that, is, when
the server is stopped).