Fix potential disordering of writes
Writing should be asynchronous, and this was achieved using a `start`
block. However, that can lead to writes becoming disordered since many
threads compete to run such code. Since we're writing into a memory BIO,
which shouldn't even block, we can fix this by moving the write into the
BIO out of the `start` block, and the scheduling just the part that does
the network interaction (reading out of the memory BIO and writing to
the underlying IO::Socket::Async).

The problem was uncovered by a sporadic Cro::HTTP test failure in the
HTTP/2 tests, which for various reasons were especially good at exposing
this issue when under load.