Partial fix for the segfault on server closing (#63)
There is a race between a possible new connection to the server
and closing the server, nullifying the context, which led to a segfault
in handle-connection on SSL_new call. Using lock to ensure
we have the context ensures we do less segfaults, however
it is not a complete fix yet.