Implement support for sslkeylogfile
The SSLKEYLOGFILE mechanism as implemented in e.g. Firefox, Chrome and Curl
listens to the SSLKEYLOGFILE environment variable and, when set to a
filename, writes the TLS session keys to that file when they are generated.
Wireshark in turn can be configured to read the keys in that file and then
transparently introspect the encrypted TLS traffic.
(Preferences -> Protocols -> TLS -> (Pre-)-Master-Secret log filename)
This implementation does not listen to an environment variable, but adds an
option `ssl-key-log-file` to the `connect` method that serves the same
purpose.
Using this option obviously fundamentally flaws the session security, so
never ever expose this option to anyone. This is solely a debugging
utility!